// admin/src/app/@core/models/pequeno-grupo-membro.model.ts
export class PequenoGrupoMembroModel {
  id: number;
  pequenoGrupoId: number;
  nome: string;
  dataNascimento: Date;
  telefone: string;
  email: string;
  cep: string;
  ruaAvenida: string;
  numero: string;
  bairro: string;
  cidade: string;
  estado: string;
  complemento: string;
  tipo: string;
  status: string;
  dataEntrada: Date;
  dataSaida: Date;
  observacao: string;
  dataCadastro: Date;
}